requirements.txt
matplotlib==3.7.1
Pillow==9.5.0
PyQt6==6.5.0
python-dotenv==1.0.0
Requests==2.29.0
